+++ Insdustrial/EnergyStorageInventory.cpp
@@ -23,7 +23,7 @@ void EnergyStorageInventory::Draw()
 	if (maxEnergy != 0)
 	{
 		functions.DrawRectangleGradient(sf::Vector2f(400, 350), sf::Vector2f(30, ((energy / (float)maxEnergy)) * -200), sf::Color::Red, sf::Color(255, 200, 0));
-		functions.PrintText(std::to_string(energy) + " / " + std::to_string(maxEnergy) + " en", sf::Vector2f(600, 200), 25, sf::Color(colorsInventory[0]));
+		functions.PrintText(std::to_string((int)energy) + " / " + std::to_string(maxEnergy) + " en", sf::Vector2f(600, 200), 25, sf::Color(colorsInventory[0]));
 	}
 }
